Crud chức vụ

+++ client/src/app/services/auth.service.ts
@@ -1,14 +1,16 @@
 import { HttpClient } from '@angular/common/http';
-import { computed, DestroyRef, effect, inject, Injectable, signal } from '@angular/core';
+import { computed, inject, Injectable, PLATFORM_ID, signal } from '@angular/core';
 import { environment } from '../../environments/environment.development';
 import { LoginRequest, LoginResponse } from '../types/login.model';
 import { catchError, Observable, tap, throwError } from 'rxjs';
+import { isPlatformBrowser } from '@angular/common';
 
 @Injectable({
   providedIn: 'root'
 })
 export class AuthService {
   private readonly http = inject(HttpClient);
+  private readonly platformId = inject(PLATFORM_ID);
   private readonly baseUrl = `${environment.apiUrl}`;
 
   private _authState = signal<LoginResponse | null>(null);
@@ -17,16 +19,18 @@ export class AuthService {
   isLoggedIn = computed(() => !!this._authState()?.accessToken);
 
   constructor() {
-    // Check if user is already logged in
-    const token = localStorage.getItem('accessToken');
-    const refreshToken = localStorage.getItem('refreshToken');
-    if (token && refreshToken) {
-      this._authState.set({
-        tokenType: 'Bearer',
-        accessToken: token,
-        expiresIn: 0,
-        refreshToken: refreshToken
-      });
+    // Only access localStorage in browser environment
+    if (isPlatformBrowser(this.platformId)) {
+      const token = localStorage.getItem('accessToken');
+      const refreshToken = localStorage.getItem('refreshToken');
+      if (token && refreshToken) {
+        this._authState.set({
+          tokenType: 'Bearer',
+          accessToken: token,
+          expiresIn: 0,
+          refreshToken: refreshToken
+        });
+      }
     }
   }
 
@@ -35,8 +39,10 @@ export class AuthService {
       .pipe(
         tap((res) => {
           this._authState.set(res);
-          localStorage.setItem('accessToken', res.accessToken);
-          localStorage.setItem('refreshToken', res.refreshToken);
+          if (isPlatformBrowser(this.platformId)) {
+            localStorage.setItem('accessToken', res.accessToken);
+            localStorage.setItem('refreshToken', res.refreshToken);
+          }
         }),
         catchError((err) => {
           console.error('Login failed', err);
@@ -47,10 +53,16 @@ export class AuthService {
 
   logout() {
     this._authState.set(null);
-    localStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.clear();
+    }
   }
 
   refreshToken() {
+    if (!isPlatformBrowser(this.platformId)) {
+      return;
+    }
+    
     const refreshToken = localStorage.getItem('refreshToken');
     return this.http.post<LoginResponse>(`${this.baseUrl}/refresh`, { refreshToken })
       .subscribe({
